pyproject.toml
[project]
name = "ai-ops"
version = "0.2.0"
description = "Multi-agent AI operations system"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "langgraph>=0.2.0",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.40.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.5.0",
    "mypy>=1.10",
]
all = [
    "anthropic>=0.40.0",
    "pytest>=8.0",
    "ruff>=0.5.0",
    "mypy>=1.10",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
